Polishing refactor with better tests
- cassandra_spec.lua doesn't have keyauth and ratelimiting tests too.
  Those are now living in `plugins/`
- cassandra_spec.lua now focuses on the Base DAO behaviour, instead of
  the entities schemas and their validation (that is
  `entities_schema_spec.lua`'s role)
- DAOs `_queries` is renamed `queries`
- `query_builder.lua` returns a boolean indicating whether or not ALLOW
  FILTERING was added to the query
