+++ spec/unit/dao/cassandra/query_builder_spec.lua
@@ -27,18 +27,21 @@ describe("Query Builder", function()
 
     describe("WHERE", function()
       it("should not allow filtering if all the queried fields are indexed", function()
-        local q = builder.select("apis", {name="mockbin"}, apis_details)
+        local q, _, needs_filtering = builder.select("apis", {name="mockbin"}, apis_details)
         assert.equal("SELECT * FROM apis WHERE name = ?", q)
+        assert.False(needs_filtering)
       end)
 
       it("should not allow filtering if all the queried fields are primary keys", function()
-        local q = builder.select("apis", {id="1"}, apis_details)
+        local q, _, needs_filtering = builder.select("apis", {id="1"}, apis_details)
         assert.equal("SELECT * FROM apis WHERE id = ?", q)
+        assert.False(needs_filtering)
       end)
 
       it("should not allow filtering if all the queried fields are primary keys or indexed", function()
-        local q = builder.select("apis", {id="1", name="mockbin"}, apis_details)
+        local q, _, needs_filtering = builder.select("apis", {id="1", name="mockbin"}, apis_details)
         assert.equal("SELECT * FROM apis WHERE name = ? AND id = ?", q)
+        assert.False(needs_filtering)
       end)
 
       it("should not allow filtering if all the queried fields are primary keys or indexed", function()
@@ -47,8 +50,9 @@ describe("Query Builder", function()
       end)
 
       it("should enable filtering when more than one indexed field is being queried", function()
-        local q = builder.select("apis", {name="mockbin", public_dns="mockbin.com"}, apis_details)
+        local q, _, needs_filtering = builder.select("apis", {name="mockbin", public_dns="mockbin.com"}, apis_details)
         assert.equal("SELECT * FROM apis WHERE name = ? AND public_dns = ? ALLOW FILTERING", q)
+        assert.True(needs_filtering)
       end)
     end)
